recatored getAge and fixed input bug
I have finally shipped a good version of the age algo that accounts
for human concept of time accurately, but the code itself could get a
lot cleaner.

As for the input bug, the month that user's submitted was 1 indexed
while js uses 0-indexed months, so it was fixed by -1 the inputted
month.