new age algo passes 3 tests
developed a new concept to count overshot days: the whole month pivot
(WMP). A WMP represents the date, with the same month as the present,
that is a perfect number of months away from the bday. This is really
neat bc you can use a WMP as the reference for both month and day count
of days live since a WMP never has to account for days.