added new hacky solution to resubtmit problem
reducer approach was deemed bad bc we don't need to track state to display
the age in screen, and the less state to keep track of the better the
code is percieved. Now, after every submit, we simple reset the display
back to its default/error phase and  then automaically re-render the age
with the new submission. ie: we rever the code to inital state (ie of
ie: as in when you first load the page (but only regarding the output
age part of the app ))